requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
astroid==1.3.6
blinker==1.4
docutils==0.12
itsdangerous==0.24
json-rpc==1.10.2
lockfile==0.10.2
logilab-common==1.0.2
pylint==1.4.4
python-bitcoinlib==0.4.0
requests==2.7.0
simplejson==3.8.0
six==1.9.0
